Pull data access from controller to store
