+++ app/services/custom_landing_page.rb
@@ -1,0 +1,4 @@
+module CustomLandingPage
+  class LandingPageConfigurationError < StandardError; end
+  class LandingPageContentNotFound < StandardError; end
+end
